函数定义和实现分离
